// src/types/contract.ts
export interface ExtractedData {
  locatario_nome?: string;
  locatario_nacionalidade?: string;
  locatario_estado_civil?: string;
  locatario_profissao?: string;
  locatario_rg?: string;
  locatario_cpf?: string;
  locatario_endereco?: string;
  locatario_cidade?: string;
  locatario_estado?: string;

  locataria_nome?: string;
  locataria_nacionalidade?: string;
  locataria_estado_civil?: string;
  locataria_profissao?: string;
  locataria_rg?: string;
  locataria_cpf?: string;
  locataria_endereco?: string;
  locataria_cidade?: string;
  locataria_estado?: string;

  locador_nome?: string;
  locador_nacionalidade?: string;
  locador_estado_civil?: string;
  locador_profissao?: string;
  locador_rg?: string;
  locador_cpf?: string;
  locador_endereco?: string;
  locador_cidade?: string;
  locador_estado?: string;

  locadora_nome?: string;
  locadora_nacionalidade?: string;
  locadora_estado_civil?: string;
  locadora_profissao?: string;
  locadora_rg?: string;
  locadora_cpf?: string;
  locadora_endereco?: string;
  locadora_cidade?: string;
  locadora_estado?: string;
}